requirements.txt
# API & serving
fastapi==0.115.0
uvicorn[standard]==0.39.0
python-multipart==0.0.12

# UI
streamlit==1.40.0

# Azure OpenAI (2.x required by llama-index-llms-openai>=0.6.26 which needs openai>=1.108.1)
# Also used for VLM vision-based image entity extraction
openai==2.32.0

# Vector store (3.0.0 required by llama-index-vector-stores-opensearch>=1.0.0)
opensearch-py[async]==3.0.0

# Graph store
neo4j==5.26.0

# Config
python-dotenv==1.0.1

# Document parsing
pypdf==6.10.2
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pillow==11.0.0
pytesseract==0.3.13

# LlamaIndex — ingestion pipeline (chunking, embeddings, vector + graph storage)
llama-index-core==0.14.16
llama-index-llms-azure-openai==0.4.2
llama-index-embeddings-azure-openai==0.4.1
# Transitive deps pinned explicitly: 0.5.0 requires core<0.14, 0.5.2 supports core<0.15
llama-index-llms-openai==0.6.26
llama-index-embeddings-openai==0.5.2
llama-index-vector-stores-opensearch==1.1.1
llama-index-graph-stores-neo4j==0.6.0
llama-index-readers-file==0.5.6

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
requests==2.32.3
